MUS-93 moved file mock
move file mock to an inner class
